Correct nigh time symbols when Sun has set
I didn't notice a bug in my nigh time symbol check to make the Moon symbols appear after Sun set. Turns out when the Sun has set, the original code was checking the setting of the FOLLOWING day, so the condition was never met, so night time symbols during night were reverting to day time. 

This update checks if Sun has set or not, and checks for appropriate rising or setting times depending if Sun has set or not.

(future update would be to swap out Sun rising and setting times with Sun dusk and dawn times, as a bright Sun icon for 8pm (here in UK in summer) looks odd when it is getting dark, meh, maybe?!  ¯\_(ツ)_/¯