Fixes day/night detection and an indentation bug
Day/night detection is used in sensor.yaml to select the appropriate
weather icon (e.g. 'sunny' vs 'night'). The previous if condition
did not work for the weather_condition_x states, when the _now state
was in a different part of the day/night cycle.
This commit should fix this issue, as long as weather_condition_x is
not too far in the future (not more than one day or night phase).

Additional bug fix:
- fix an indentation bug in weatherman.yaml