Update sensor.yaml - for new Weather Forecast Service
With the new HA Weather Forecast service to get the weather, Met.no no longer has the option for an "hourly" entity. new users of this sensor and project won't be able to get the Met.no Hourly entity.

This update adds in the "service: weather.get_forecast" call to get the hourly result. This is then added to the sensor using variable "weather_valhalla_hourly".

It assumes your met.no weather entity is called "weather.valhalla" - adjust accordingly in the action entity_id.